#381 refactor for pidgin body filter
