+++ src/f10_world/pidgin_agg.py
